+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable user-facing changes to mssql-django are documented in this file.
 
 ## [Unreleased]
 
+### Fixed
+
+- Allowed mssql-python connections to start without the unixODBC driver
+  manager when pyodbc is not selected ([#621]).
+
 ## [2.0.0] - 2026-09-18
 
 ### Added
@@ -160,3 +165,4 @@ All notable user-facing changes to mssql-django are documented in this file.
 [#605]: https://github.com/microsoft/mssql-django/pull/605
 [#609]: https://github.com/microsoft/mssql-django/pull/609
 [#613]: https://github.com/microsoft/mssql-django/pull/613
+[#621]: https://github.com/microsoft/mssql-django/pull/621